+++ README.md
@@ -1,11 +1,6 @@
 # AllDutchZipcodeFinder
 
-Most recent release: https://github.com/LegeBeker/AllDutchZipcodeFinder/releases/tag/0.1.1
+**Most recent release:** *https://github.com/LegeBeker/AllDutchZipcodeFinder/releases/tag/0.3.0*
 
-Find all dutch zipcodes between 4 digit numbers
-
-This is a program I wrote in my free time to learn more about Python.
-
-I am not giving a guarantee that all zipcodes are present.
-
-I used this create a database of all possible zipcodes: https://github.com/jorritb/dutch-postal-codes
+This is a program I wrote in my freetime to find all existing dutch zipcodes between 2 numbers containing 4 digits, and generate a file on your desktop with the result.
+I am not giving a guarantee that all zipcodes are present. I used this create a database of all possible zipcodes: *https://github.com/jorritb/dutch-postal-codes*
